Generalized main tenant
